requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
numpy==1.24.3
torch==2.0.1
transformers==4.33.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
httpx==0.25.0